requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
fireworks-ai==0.15.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.45
langgraph==0.3.11
langgraph-checkpoint==2.0.20
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.15
msgpack==1.1.0
numpy==2.2.4
openai==1.66.3
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wsproto==1.2.0
zstandard==0.23.0
